+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Thu, 07 Jul 2022 21:24:07 GMT and should not be manually modified.
+This log was last generated on Fri, 08 Jul 2022 21:23:36 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.3
+
+Fri, 08 Jul 2022 21:23:36 GMT
+
+### Patches
+
+- Move used tokens types to tokens package (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.7.32
+- Bump @fluentui-react-native/text to v0.11.28
+- Bump @fluentui-react-native/tokens to v0.15.1
 
 ## 0.12.2